+++ src/lib.rs
@@ -2,5 +2,6 @@ pub use wasm_bindgen_rayon::init_thread_pool;
 
 pub mod config;
 pub mod model;
+pub mod utils;
 pub mod moshi_worker;
 pub use moshi_worker::MoshiASRDecoder;
